// Sample.TMLink.TypeScript/src/Protos/cyg_tml_api_v1.ts
export enum CommandType {
  Non = 0,
  /** GetGaugeInfo - requests info about the Gauge (serial, model, variant, battery level etc). */
  GetGaugeInfo = 1,
  /** CancelRecordTransfer - send to cancel a Record or BScan transfer from the gauge. */
  CancelRecordTransfer = 2,
  /** GetRecordList - requests a list of Records on the gauge. */
  GetRecordList = 3,
  /** GetRecord - requests a Record from the gauge. */
  GetRecord = 4,
  /** GetRecordPoint - requests a Record's measurement point (No AScan data). */
  GetRecordPoint = 5,
  /** GetRecordPointAScan - requests a Record's measurement point (With AScan data). */
  GetRecordPointAScan = 6,
  /** DeleteRecord - requests a Record is deleted from the gauge. */
  DeleteRecord = 7,
  /** DeleteAllRecords - requests all Records are deleted from the gauge. */
  DeleteAllRecords = 8,
  /** NewRecord - sends a New blank Record to the Gauge. */
  NewRecord = 9,
  /** AddRecordPoints - sends a block of 1..10 blank Points and adds to the new Record on the Gauge. */
  AddRecordPoints = 10,
  /** GetBScanList - requests a list of BScans on the gauge. */
  GetBScanList = 11,
  /** GetBScan - requests a BScan from the gauge. */
  GetBScan = 12,
  /** GetBScanPoint - requests a BScan's measurement point (No AScan data). */
  GetBScanPoint = 13,
  /** GetBScanPointAScan - requests a BScan's measurement point (With AScan data). */
  GetBScanPointAScan = 14,
  /** DeleteBScan - requests a BScan is deleted from the gauge. */
  DeleteBScan = 15,
  /** DeleteAllBScans - requests all BScans are deleted from the gauge. */
  DeleteAllBScans = 16,
  UNRECOGNIZED = -1,
}
// ...
export enum ErrorCodes {
  /** Success - Success, no error. */
  Success = 0,
  /** RecordNotFound - The Record was not found. */
  RecordNotFound = 1,
  /** BScanNotFound - The BScan was not found. */
  BScanNotFound = 2,
  /** InvalidParameter - An invalid parameter was received in a command. */
  InvalidParameter = 3,
  /** TransferNotStarted - Attempt to get Points before Header. */
  TransferNotStarted = 4,
  /** GzUncompressError - Gzip uncompress error. */
  GzUncompressError = 5,
  /** ProtoUnpackError - Protobuf unpack error. */
  ProtoUnpackError = 6,
  /** MemoryError - Gauge memory allocation error. */
  MemoryError = 7,
  /** RecordSaveError - Failed to save Record data. */
  RecordSaveError = 8,
  /** RecordNameError - Record name is invalid. */
  RecordNameError = 9,
  /** RecordZeroPoints - Attempt to created a record with zero points. */
  RecordZeroPoints = 10,
  /** RecordPointNumError - Attempt to created a record with too many points. */
  RecordPointNumError = 11,
  /** TransferCancelled - Record or BScan transfer cancelled. */
  TransferCancelled = 12,
  UNRECOGNIZED = -1,
}
// ...
export enum LiveMeasurementType {
  Live = 0,
  Frozen = 1,
  UNRECOGNIZED = -1,
}
// ...
export enum ProbeType {
  /** Single - Single crystal probe, direct contact. */
  Single = 0,
  /** Dual - Dual crystal probe, with delay line. */
  Dual = 1,
  UNRECOGNIZED = -1,
}
// ...
export interface NotifyMessage {
  /** The command that triggered this message. */
  commandType: CommandType;
  /** Error code. */
  errorCode: ErrorCodes;
  /** True if data for the client to read. */
  readDataAvailable: boolean;
}

/**
 * ---------------------------------------------------------------------------------------------------
 * ## Command contains the Command send from the Client to the gauge.
 * Used for the BLE Service 'WriteCommand' Write characteristic.
 * Command messages are always gzip compressed.
 */
export interface Command {
  /** The command type being sent in the message. */
  commandType: CommandType;
  /** Sets Gauge date-time when non-zero is sent. */
  timestamp:
    | Date
    | undefined;
  /** Optional name of a Record or BScan. (30 chars max). */
  name: string;
  /** Optional newRecord message. */
  newRecord:
    | Command_NewRecord
    | undefined;
  /** Optional addRecordPoints message. */
  addRecordPoints: Command_AddRecordPoints | undefined;
}

/**
 * ------------------------------------------------------------------------------------------------
 * Used to create a new Record in the gauge.
 */
export interface Command_NewRecord {
  /** record name or ID (30 chars max). */
  name: string;
  /** TML Key or ID  (30 chars max). */
  key: string;
  /** Linear or Grid */
  recordType: RecordType;
  /** Grid Pattern (if Grid type) */
  gridPattern: GridPattern;
  /** Units of Measurement. */
  uom: UOM;
  /** Linear or Grid X Dimension 1..N. */
  numColsX: number;
  /** Grid Y coordinate 1..N. */
  numRowsY: number;
}

/**
 * ------------------------------------------------------------------------------------------------
 * Used to add blank points to a new record.
 * Send in blocks of 1..10 points.
 * If more than 10 points are sent, the message will be discarded and a NAK response returned.
 */
export interface Command_AddRecordPoints {
  /** record name or ID (30 chars max). */
  name: string;
  /** Array of 1..10 measurement points per message. */
  mpoints: Command_AddRecordPoints_MPoint[];
}
// ...
export interface Message {
  /** The command that triggered this message. */
  commandType: CommandType;
  /** An optional GaugeInfo message. */
  gaugeInfo:
    | Message_GaugeInfo
    | undefined;
  /** An optional Record List. */
  recordList:
    | Message_RecordList
    | undefined;
  /** An optional Record. */
  record:
    | Message_Record
    | undefined;
  /** An optional RecordPoint. */
  recordPoint:
    | Message_RecordPoint
    | undefined;
  /** An optional BScan List. */
  bscanList:
    | Message_BScanList
    | undefined;
  /** An optional BScan. */
  bscan:
    | Message_BScan
    | undefined;
  /** An optional BScanPoint. */
  bscanPoint: Message_BScanPoint | undefined;
}
// ...
export interface Message_Record {
  /** RecordID. */
  recordID: number;
  /** record name or ID (30 chars max). */
  name: string;
  /** location name or ID (16 chars max). */
  location: string;
  /** surveyor name or ID (16 chars max). */
  surveyor: string;
  /** TML Key or ID (30 chars max). */
  key: string;
  /** timestamp when record file created. */
  created:
    | Date
    | undefined;
  /** timestamp when record file last saved. */
  updated: Date | undefined;
  recordType: RecordType;
  numPointsRequired: number;
  /** The number of RecordPoint messages to read. */
  numPointsTaken: number;
}
// ...
export interface NotifyLiveMeasurement {
  /** Informs of the Live Measurement type. */
  liveMeasurementType: LiveMeasurementType;
  /** If liveMeasurementType == Live these items are included.. */
  index: number;
  /** Status Bits x 5 [DEEPCOAT, INCHES, FROZEN, STABLE, VALID](LSB). */
  statusBits: number;
  /** UT Measurement Mode; SE/EE/ME. */
  UTMode: UTMode;
  /** Thickness value in UOM. */
  thickness: number;
  /** Velocity of sound in UOM. */
  velocity: number;
  /** Gain set in dB */
  gaindB: number;
  /** Surface temp in C if High Temp mode, else zero. */
  surfaceTemp: number;
  /** Battery level in percent. */
  batteryLevel: number;
}

/**
 * ------------------------------------------------------------------------------------------------
 * ## Contains a single Frozen Live Measurement.
 * FrozenLiveMeasurement messages are always gzip compressed.
 */
export interface FrozenLiveMeasurement {
  /** Measurement index number. */
  index: number;
  /** Gauge serial number. */
  serialNumber: number;
  /** Gauge ID number. */
  gaugeUD: number;
  /** Battery level in percent. */
  batteryLevel: number;
  /** True when valid measurement. */
  validMeasurement: boolean;
  /** True when stable measurement. */
  stableMeasurement: boolean;
  /** Units of Measurement. */
  uom: UOM;
  /** Thickness value in UOM. */
  thickness: number;
  /** Velocity of sound in UOM. */
  velocity: number;
  /** Probe type; Single, Dual. */
  probeType: ProbeType;
  /** UT Measurement Mode; SE/EE/ME. */
  UTMode: UTMode;
  /** Gain set in dB */
  gaindB: number;
  /** Surface temp In C if High Temp mode, else zero. */
  surfaceTemp: number;
  /** AScan data. */
  ascan: AScan | undefined;
}
// ...
export const Command: MessageFns<Command> = {
  encode(message: Command, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.commandType !== 0) {
      writer.uint32(8).int32(message.commandType);
    }
    if (message.timestamp !== undefined) {
      Timestamp.encode(toTimestamp(message.timestamp), writer.uint32(18).fork()).join();
    }
    if (message.name !== "") {
      writer.uint32(26).string(message.name);
    }
    if (message.newRecord !== undefined) {
      Command_NewRecord.encode(message.newRecord, writer.uint32(34).fork()).join();
    }
    if (message.addRecordPoints !== undefined) {
      Command_AddRecordPoints.encode(message.addRecordPoints, writer.uint32(42).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): Command {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseCommand();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.commandType = reader.int32() as any;
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.timestamp = fromTimestamp(Timestamp.decode(reader, reader.uint32()));
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.name = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.newRecord = Command_NewRecord.decode(reader, reader.uint32());
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.addRecordPoints = Command_AddRecordPoints.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): Command {
    return {
      commandType: isSet(object.commandType) ? commandTypeFromJSON(object.commandType) : 0,
      timestamp: isSet(object.timestamp) ? fromJsonTimestamp(object.timestamp) : undefined,
      name: isSet(object.name) ? globalThis.String(object.name) : "",
      newRecord: isSet(object.newRecord) ? Command_NewRecord.fromJSON(object.newRecord) : undefined,
      addRecordPoints: isSet(object.addRecordPoints)
        ? Command_AddRecordPoints.fromJSON(object.addRecordPoints)
        : undefined,
    };
  },

  toJSON(message: Command): unknown {
    const obj: any = {};
    if (message.commandType !== 0) {
      obj.commandType = commandTypeToJSON(message.commandType);
    }
    if (message.timestamp !== undefined) {
      obj.timestamp = message.timestamp.toISOString();
    }
    if (message.name !== "") {
      obj.name = message.name;
    }
    if (message.newRecord !== undefined) {
      obj.newRecord = Command_NewRecord.toJSON(message.newRecord);
    }
    if (message.addRecordPoints !== undefined) {
      obj.addRecordPoints = Command_AddRecordPoints.toJSON(message.addRecordPoints);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<Command>, I>>(base?: I): Command {
    return Command.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<Command>, I>>(object: I): Command {
    const message = createBaseCommand();
    message.commandType = object.commandType ?? 0;
    message.timestamp = object.timestamp ?? undefined;
    message.name = object.name ?? "";
    message.newRecord = (object.newRecord !== undefined && object.newRecord !== null)
      ? Command_NewRecord.fromPartial(object.newRecord)
      : undefined;
    message.addRecordPoints = (object.addRecordPoints !== undefined && object.addRecordPoints !== null)
      ? Command_AddRecordPoints.fromPartial(object.addRecordPoints)
      : undefined;
    return message;
  },
};
// ...
export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}